requirements.txt
httpx
mcp
langgraph
langgraph-checkpoint-sqlite
langchain
langchain-google-genai
langchain-openai
python-dotenv

# Core ML/DL libraries
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0
sentencepiece>=0.1.99
tokenizers>=0.15.0

# Training utilities
tqdm>=4.66.0
wandb>=0.16.0
tensorboard>=2.15.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
pyarrow>=14.0.0  # For reading parquet files from HuggingFace

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# API & Integration
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
rich>=13.7.0
loguru>=0.7.2

# Tier 2 & 3: Memory & RAG
# Tier 2: SQLite already included with Python
chromadb>=1.4.0  # Tier 3: Vector database
sentence-transformers>=5.0.0  # Tier 3: Embeddings
tiktoken>=0.5.0  # Tier 1: Token counting
